Complete externalization of database configuration
This removes the code we had that would attempt to configure a mysql database
from scratch. That code was problematic for a number of reasons (assumed an
account name of 'root', no argument for passwords, etc), but ultimately there's
enough complexity in setting up a database that attempting to do that as part
of this module is an endeavour fraught with peril. Better to externalize that
so that the user can use another module, or manual action, or whatever else.